fixed instanced draw layer limitation
drawInstanced uses a ImageArray were the characterAnimator.shader sets the right layer on the fragment shader. before testing was under 2048 layers which is a limitation if each layer only contained the relevant draw for a single sprite cutout. now we use a 2k texture on each layer. this brings the total pixel representation way up from 2048*128*128 to 2048^3. this involves a uv lookup into the texture atlas which means the way quads are being first put into the imageArray is now more important.